backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.26.0
scikit-learn>=1.4.0
pydantic>=2.6.0 